Dynamically allocate the nvs_path for plt_tx_bip.
In plt_tx_bip, the nvs_path was originally allocated on the stack
and then passed as a parameter for the calib_valid_handler netlink
callback. However, the callback will not be processed until the
plt_tx_bip handler returns to __handle_cmd() and the responses
all processed by nl_recvmsgs at the end of the function. Other
functions are called between the handler and the nl_recvmsgs that
were overwriting the stack on my platform. By migrating the path
to a dynamically allocated string, the path is not corrupted and
the calibration process to proceed on my platform.